+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 1. dialog component
 2. "server subscribe page"
 3. use zustand to replace context
+4. try to use pipe between agent and rust/ts, but failed
 
 ### 0.4.4
 1. better module config page